Add uglify.output Option in Build
`uglify` option in `build` now accepts `uglify.output` consistent with the Uglify2 API.

This change does NOT change existing behaviour as the inconsistent `uglify.beautify` option is still valid. Internally `uglify.output` is copied into `uglify.beautify`, when the latter is not specified. If `uglify.beautify` is specified, then `uglify.output` is ignored.