Use terser to support minification of more permissive babel presets
Currently, only ES5 javascript can be minified, due to the reliance on
uglify-js. With terser (a maintained fork of the harmony branch of
UglifyJS2) it is possible to support a broader range of javascript
syntax.

The API changed between uglify-js@2 and uglify-js@3 / terser, but the
translation to the new API was fairly straightforward. It hews more
closely to the original code than #815, and consequently I think it
avoids regressing any features (for instance, the ability to include
sourceContents in source maps or to control the comment stripping is
retained).

As browsers continue to support more features beyond ES5, and
because babel recommends use of the 'env' preset, it will become
increasingly likely that users' babel transpilation settings will cause
errors when trying to minify with jspm and friends as long as it uses
uglify-js.

refs #815, #726